requirements.txt
aiogram>=3.8.0
pydantic>=2.8.0
SQLAlchemy==2.0.31
python-dotenv==1.0.1
Pillow==10.4.0
skia-python
Jinja2==3.1.4